refactor: separate process responses and types
